requirements.txt
Flask==1.1.1
imutils==0.5.4
matplotlib==3.1.2
numpy==1.17.4
opencv_python==4.5.3.56
Pillow==8.4.0
protobuf==3.19.1
PySimpleGUI==4.55.1
tensorflow==2.7.0
